+++ lib/helper/image_generation_helper.dart
@@ -1,0 +1,101 @@
+import 'dart:io';
+import 'package:get/get.dart';
+import 'package:matrix_ai/controller/generation_controller.dart';
+import 'package:matrix_ai/controller/image_generation_controller.dart';
+import 'package:matrix_ai/controller/settings_controller.dart';
+import 'package:matrix_ai/controller/subscription_controller.dart';
+import 'package:matrix_ai/view/base/ads_dialog.dart';
+import 'package:matrix_ai/view/screens/prompt_details/prompt_details.dart';
+import 'package:matrix_ai/helper/navigation.dart';
+import 'package:matrix_ai/common/snackbar.dart';
+import 'package:matrix_ai/view/base/rate_us_sheet.dart';
+import 'package:matrix_ai/view/screens/subscription/subscription.dart';
+
+class ImageGenerationHelper {
+  static Future<void> handleTap(Function(String) handleImageGeneration) async {
+    final settings = SettingsController.find;
+    final text = settings.promptController.text;
+
+    if (_isPromptEmpty(text)) {
+      showToast('please_enter_prompt'.tr);
+    } else if (_hasOffensiveWords()) {
+      showToast('please_remove_offensive_words'.tr);
+    } else {
+      handleImageGeneration(text);
+    }
+  }
+
+  static bool _isPromptEmpty(String text) => text.isEmpty;
+
+  static bool _hasOffensiveWords() => SettingsController.find.hasOffensiveWords;
+
+  static Future<void> handleImageGeneration(String text,
+      {required Function(String, {bool showAds}) generateImage}) async {
+    if (Platform.isAndroid || SubscriptionController.find.isPro) {
+      await _handleProOrAndroidUser(text, generateImage: generateImage);
+    } else {
+      _handleFreeUser(text, generateImage: generateImage);
+    }
+  }
+
+  static Future<void> _handleProOrAndroidUser(String text,
+      {required Function(String, {bool showAds}) generateImage}) async {
+    if (GenerationController.find.proUserLimitExceeded) {
+      bool hasShowedFreeLimitDialog =
+          await ImageGenerationController.find.hasShowedFreeLimitDialog();
+      if (hasShowedFreeLimitDialog) {
+        generateImage(text, showAds: false);
+      }
+    } else {
+      generateImage(text);
+    }
+  }
+
+  static void _handleFreeUser(String text,
+      {required Function(String, {bool showAds}) generateImage}) {
+    if (_isProModel()) {
+      showPremiumSheet();
+    } else {
+      _handleFreeUserGeneration(text, generateImage: generateImage);
+    }
+  }
+
+  static Future<void> _handleFreeUserGeneration(String text,
+      {required Function(String, {bool showAds}) generateImage}) async {
+    bool hasShowedFreeLimitDialog =
+        await ImageGenerationController.find.hasShowedFreeLimitDialog();
+    if (hasShowedFreeLimitDialog) {
+      generateImage(text, showAds: false);
+    } else {
+      showAdsDialog(
+        onWatchAdPressed: () {
+          pop();
+          generateImage(text);
+        },
+      );
+    }
+  }
+
+  static bool _isProModel() =>
+      SettingsController.find.configModel.selectedModel?.premium ?? false;
+
+  static void generateImage(String text, {bool showAds = true, int? seed}) {
+    ImageGenerationController.find
+        .generateImages(text, seed: seed, showAds: showAds)
+        .then(
+      (response) {
+        if (response != null) {
+          bool fromRegenerate = seed != null;
+          if (fromRegenerate) {
+            pop();
+          }
+          launchScreen(PromptDetailScreen(response: response),
+              replace: fromRegenerate);
+          Future.delayed(const Duration(seconds: 2), () {
+            showConditionalRateUsDialog();
+          });
+        }
+      },
+    );
+  }
+}